+++ Geoscape/RecruitInfo/changelog.md
@@ -1,8 +1,9 @@
 Changelog of Recruit Info, a Phoenix Point Mod by Sheepy
 
-# Version 4, in development
+# Version 4, 2020-11-02
 
 * Update to work with Phoenix Point 1.7.3.  Mod is not backward compatible with pre-1.7 game versions.
+* Fix: Medkit numbers are now properly interpolated in the description.
 
 # Version 3, 2020-06-04